Update metaphlan_sbatch.sh
Updated num_jobs to 50 as default. Also added shell commands for copying over all snakemake related files including logs so that a particular project parameter specifications, conda environment used (program versions for publications) are available and not confusing to follow. Also added the sbatch command job name so that it is easier to see in the squeue list.